[dvsim/syn/lint] Add switches to selectively sanitize reports
We have three different use cases for the detailed, non-summary reports:

1) local command line reporting
2) email reports
3) published reports

For 1) we would always like to get the maximum amount of details, and
since it is only generated and consumed locally, this is ok to do.

For 2) and 3) we should be able to selectively turn on sanitization,
which will prevent detailed messages from being sent out via mail or
from being published to the web (we are not allowed to publish detailed
messages for certain tools due to license and NDA restrictions).

This patch adds such switches, and configures our DC synthesis flow to
only publish a sanitized report to the web.

Signed-off-by: Michael Schaffner <msf@google.com>