[prim_alert_sender] Clear ping req when sigint condition resolves
For the in-band alert channel reset feature, we intend to leverage sigint
errors on the ping/ack lines to re-initialize the sender FSM and clear any
pending ping requests on the sender side.

This change makes sure that upon deassertion of a sigint condition, we
actually clear the ping request set-reg.

Signed-off-by: Michael Schaffner <msf@google.com>