[prim_alert_sender] Simplify sender and clear ping req upon sigint
This simplifies the sender state machine and removes the explicit SigInt
state. It also simplifies the response behavior which so far toggled the
alert lines during a signal integrity error. The new behavior is that
the alert diff pair is just pulled to zero in case of a signal integrity
alert, thereby simplifying the protocol.

Further, for the in-band alert channel reset feature, we intend to leverage
sigint errors on the ping/ack lines to re-initialize the sender FSM and
clear any pending ping requests on the sender side.

This change makes sure that upon deassertion of a sigint condition, we
actually clear the ping request set-reg.

Signed-off-by: Michael Schaffner <msf@google.com>