[prim_alert_sender] Simplify sender and remove SigInt state
This simplifies the sender state machine and removes the explicit SigInt
state. It also simplifies the response behavior which so far toggled the
alert lines during a signal integrity error. The new behavior is that
the alert diff pair is just pulled to zero in case of a signal integrity
alert, thereby simplifying the protocol.

Signed-off-by: Michael Schaffner <msf@google.com>