[rv_dm,lc_ctrl,darjeeling] Convert DMI to real TL-UL interface
Change DMI implementation to TL-UL, and rework rv_dm and lc_ctrl to
match. Ensure rv_dm continues to respond with zeroes when the DM is not
present / inaccessible.

Rewire JTAG to place lc_ctrl and rv_dm on the same TAP, and use
dedicated DIOs for JTAG at the top. Generate a RAL model for the new
TL-UL interconnect, and use it in DV. Also fix up macro names in
software for the new lc_ctrl named register interfaces.

rv_dm is located at address 0 in the map. This is required by the RISC-V
Debug specification, as it is the first Debug Module. If other Debug
Modules are to be used, they could be placed elsewhere on the TL-UL
network, and the nextdm CSR for the first DM would point to the base
address of the next.

lc_ctrl is currently located at 0x20000 in the map. This may be changed
to match data sheet and/or flexibility preferences. This address
currently must be hard-coded (or chained to the number of DMI abits), as
there is no method specified for probing this address. Only the IDCODE
and what can be read from the DTMCS register are currently usable keys
for finding out the DMI addresses. Alternative mechanisms are left for
future discussions.

Signed-off-by: Alexander Williams <awill@opentitan.org>
Co-authored-by: Michael Schaffner <msf@opentitan.org>