[sysrst_ctrl] Rework IRQ sync mechanism
This reworks the IRQ event synchronization in preparation for
converting the IRQ to status type, see #21832.

In particular, it moves tracking of the IRQs into the bus domain
so that a status interrupt can be created more readily.

This is done by:
1) moving the collated IRQ status regs into the bus domain
2) creating a similar IRQ synchronization module as in adc_ctrl

Note that the wakeup request is still created and tracked in the AON
domain as before.

Signed-off-by: Michael Schaffner <msf@opentitan.org>